requirements.txt
PyYAML
Flask
pyopenssl
wheel
gunicorn
textdistance[extras]
transliterate
dateparser
bravado-core